third_party/blink/web_tests/flag-specific/highdpi/fast/forms/color/color-picker-events-coordinates-expected.txt

This is a testharness.js-based test.
FAIL Color picker: Mouse and touch events should produce the same values. assert_equals: values should be the same for both touch and mouse events expected "#3b5763" but got "#718c98"
Harness: the test ran to completion.
